refactor: diagnostics
